Explain the why of SQS to communicate across regions
Since we use Kafka for all sorts of asynchronous communication, I was
curious to understand the reasoning for choosing SQS in this particular
use case. Even for the use cases in which a message queue is more
appropriate, like this one, we often use Kafka anyway because we have
very good tooling and monitoring around it.