Edit dataset_series.md and contracts.md (#168)
* Improve the wording of snapshots datasets

In this way it‘s consistent with attribute history datasets

* Fix broken link to Itaipu contracts document

* Replace Github link to produce-to-etl! with codesearch

Rationale: Github links to a specific location in the source code that
are not also tied to a specific commit are fragile; Codesearch seems a
better solution and it‘s owned by Nubank, too. Downsides: more
laborious and it works only with an active VPN connection.

* Add remark about entity history datasets usage

* Correct reference to DatasetSeriesAttribute.transform field

* Add structure to “Dealing with versions“ section

I tried to improve the glanceability of the section. The commit is a
bit more noisy than strictly needed because of text reflows and
nesting.

PS: The part describing the renaming is still a bit confusing to me,
especially if we consider the `Final steps` question.

* Remove isPrimaryKey example declaration

Rationale:
- Simple enough
- Shown elsewhere, too
- It disrputs the reading flow. In that point, what we really care
  about is the deduplication logic.

* Update etl_users/dataset_series.md

As per reviewer’s suggestion

Co-Authored-By: Antoine Chesnais <antoine.chesnais@nubank.com.br>